scene graph based light probe
