+++ examples/webgl_lights_lightprobe.html
@@ -104,7 +104,7 @@
 				scene.add( ambientLight );
 
 				// probe
-				lightProbe = new THREE.LightProbe( 0xffffff, API.lightProbeIntensity );
+				lightProbe = new THREE.DiffuseLightProbe( 0xffffff, API.lightProbeIntensity );
 				scene.add( lightProbe );
 
 				// light
@@ -137,7 +137,7 @@
 					//var geometry = new THREE.TorusKnotBufferGeometry( 4, 1.5, 256, 32, 2, 3 );
 
 					var material = new THREE.MeshStandardMaterial( {
-						color: 0xffffff, 
+						color: 0xffffff,
 						metalness: 0,
 						roughness: 0,
 						envMap: cubeTexture,
